+++ src/main/java/com/project/acre/util/SHA256.java
@@ -1,0 +1,20 @@
+package com.project.acre.util;
+
+import java.security.MessageDigest;
+import java.security.NoSuchAlgorithmException;
+import java.util.Base64;
+
+public class SHA256 {
+    private static final SHA256 SHA256 = new SHA256();
+
+    public static SHA256 getInstance() {
+        return SHA256;
+    }
+
+    public String encrypt(String password) throws NoSuchAlgorithmException {
+        MessageDigest md = MessageDigest.getInstance("SHA-256");
+        md.update(password.getBytes());
+        byte[] digest = md.digest();
+        return Base64.getEncoder().encodeToString(digest);
+    }
+}
